backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
python-dotenv
supabase
openai
google-genai
requests
PyPDF2
python-docx
aiofiles
numpy
sentence-transformers
PyJWT
passlib[bcrypt]

# ML/AI Model packages
torch
transformers
pandas
matplotlib
scikit-learn
scipy           